% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRelFreqDist.R
\name{calculateRelFreqDists}
\alias{calculateRelFreqDists}
\title{Calculate Relative Frequence Distributions}
\usage{
calculateRelFreqDists(n_grams)
}
\arguments{
\item{n_grams}{tokens-object comprising all features for each text}
}
\description{
This function calculates relative frequency distributions for every text
in the corpus
}
\examples{
calculateRelFreqDists()
}
\keyword{distribution}
\keyword{frequency}
\keyword{relative}
