% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateZscores.R
\name{calculateZscores}
\alias{calculateZscores}
\title{Calculate Zscores}
\usage{
calculateZscores(freq_dist)
}
\arguments{
\item{freq_dist}{matrix comprising the frequency distribution for the
surviving features for each text}
}
\description{
This function calulates zscores from relative frequency distributions.
}
\examples{
calculateZscores()
}
\keyword{zscores}
