% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySVM.R
\name{classifySVM}
\alias{classifySVM}
\title{Classify Texts with SVM}
\usage{
classifySVM(path_zscores, metadata, metadata_col, col_file_name,
  seed = 100)
}
\arguments{
\item{path_zscores}{path to zscore matrix}

\item{metadata}{metadata dataframe}

\item{metadata_col}{name of column of metadata dataframe which should be
used for the classification}

\item{col_file_name}{name of column that comprises the texts file names}

\item{seed}{set seed for classification, default is 100}
}
\description{
This function implements a SVM classification.
}
\examples{
classifySVM()
}
\keyword{SVM}
\keyword{classification,}
