% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFreqDists.R
\name{combineFreqDists}
\alias{combineFreqDists}
\title{Combine Frequency Distributions}
\usage{
combineFreqDists(freq_dists, sorted_features, filenames)
}
\arguments{
\item{freq_dists}{list of frequency distributions}

\item{sorted_features}{vector containing the sorted overall features}

\item{filenames}{vector containing the file names of all texts in the corpus}
}
\description{
This function combines all frequency distributions of all texts.
}
\examples{
combineFreqDists()
}
\keyword{distribution}
\keyword{frequency}
