% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDistanceTable.R
\name{createDistanceTable}
\alias{createDistanceTable}
\title{Create a Distance Table}
\usage{
createDistanceTable(path_to_corpus, freq_dist, n_gram_size = 1,
  culling_level = 20, cut_off = 500, zscores_transformation = NULL,
  distance_measure = "burrows-delta")
}
\arguments{
\item{path_to_corpus}{path to the corpus directory}

\item{freq_dist}{frequency distribution matrix}

\item{n_gram_size}{size of ngrams - possible values: 1 & 2; default is 1}

\item{culling_level}{gives the opportunity to exclude sparse words: a
culling level of 20 means that only words which occur in at least 20% of all
texts in the corpus are considered in the analysis, default = 20}

\item{cut_off}{number of features which will be considered for the analysis,
default = 500}

\item{zscores_transformation}{declares whether the zscores should be normalised or
ternarised, default = NULL}

\item{distance_measures}{determines the distance measure which is used for the
calculation of the distance table, default = "burrows-delta"}
}
\description{
This function creates a distance table by combining several functions
}
\examples{
createDistanceTable()
}
\keyword{distance}
\keyword{table}
