% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFreqDist.R
\name{createFreqTable}
\alias{createFreqTable}
\title{Create a Frequency Table}
\usage{
createFreqTable(path_to_corpus, remove_whitespaces = TRUE,
  lowercase = TRUE, remove_punctuation = TRUE, n_gram_size = 1)
}
\arguments{
\item{path_to_corpus}{path to the corpus directory}

\item{lowercase}{Should tokens be turned into lowercase? Default is TRUE}

\item{remove_punctuation}{Should punctuation marks be removed? Default is TRUE}

\item{n_gram_size}{size of ngrams - possible values: 1 & 2; default is 1}

\item{remove_whitespace}{Should white-spaces be removed? Default is TRUE}
}
\description{
This function creates a distance table by combining several functions
}
\examples{
createFreqTable()
}
\keyword{distance}
\keyword{table}
