% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLinksNodes.R
\name{createLinksNodes}
\alias{createLinksNodes}
\title{Create Links and Nodes}
\usage{
createLinksNodes(path_distance_matrix, nearest_neighbours = TRUE,
  num_neighbours = 6, cut_off = FALSE, percentage = 5)
}
\arguments{
\item{path_distance_matrix}{path to the distance matrix}

\item{nearest_neighbours}{Should the nearest neighbour method be used?
Default is TRUE}

\item{num_neighbours}{Number of nearest neighbours which will be used in
the analysis. Default is 6}

\item{cut_off}{Should the percental cut-off method be used? Default is FALSE}

\item{percentage}{Percentage of links which should be used for the network.
Default is 5}
}
\description{
This creates
}
\examples{
createLinksNodes()
}
\keyword{links,}
\keyword{nodes}
