% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetaSubset.R
\name{createMetaSubsets}
\alias{createMetaSubsets}
\title{Create Subset}
\usage{
createMetaSubsets(metadata, subset_column, path_results)
}
\arguments{
\item{metadata}{metadata dataframe}

\item{subset_column}{column featuring categories used for the subsetting}

\item{path_results}{path to directory where results should be saved}
}
\description{
This function creates subsets from metadata tables by detecting unique
values and using these values to split up the dataset.
}
\details{
To enable the combination with zscores and/or distance tables, the metadata
tables have to feature a column featuring the respective texts filename.
}
\examples{
createSubsets()
}
\keyword{subsets}
