% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{createNetwork}
\alias{createNetwork}
\title{Create a Network}
\usage{
createNetwork(list_of_links_and_nodes, metadata, col_file_name,
  metadata_col)
}
\arguments{
\item{list_of_links_and_nodes}{object containing networkD3 links and nodes
and parameter settings}

\item{metadata}{metadata dataframe}

\item{col_file_name}{name of column that comprises the texts file name}

\item{metadata_col}{name of column of metadata dataframe which should be
used for the colouring of the nodes}
}
\description{
This function creates a network visualisation.
}
\examples{
createNetwork()
}
\keyword{network}
\keyword{visualisation}
