% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutMFF.R
\name{cutMFF}
\alias{cutMFF}
\title{Limit Most Frequent Features}
\usage{
cutMFF(freq_dist, cut_off)
}
\arguments{
\item{freq_dist}{frequency distribution matrix}

\item{cut_off}{cut-off for MFF}
}
\description{
This function limits the features to the designated cut-off.
}
\examples{
cutMFF()
}
\keyword{features}
\keyword{frequent}
\keyword{most}
