% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSpecifiedFeatures.R
\name{deleteSpecifiedFeatures}
\alias{deleteSpecifiedFeatures}
\title{Delete Specified Features}
\usage{
deleteSpecifiedFeatures(input_df, features_to_be_deleted)
}
\arguments{
\item{input_df}{frequency distribution matrix}

\item{features_to_be_deleted}{list of tokens which should be deleted}
}
\description{
This function deletes columns from a matrix if the parameter list includes
a column's name.
}
\examples{
deleteSpecifiedFeatures()
}
\keyword{delete}
\keyword{features}
