% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFilenames.R
\name{extractFilenames}
\alias{extractFilenames}
\title{Extract Filenames}
\usage{
extractFilenames(path_to_corpus)
}
\arguments{
\item{path_to_corpus}{path to the corpus directory}
}
\description{
This function extracts all filenames of a corpus for further use.
}
\examples{
extractFilenames()
}
\keyword{filenames}
