% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSubsets.R
\name{filterDistSubsets}
\alias{filterDistSubsets}
\title{Filter for Subset}
\usage{
filterDistSubsets(path_distance_matrix, file_name_col, path_results)
}
\arguments{
\item{path_distance_matrix}{path to the distance matrix}

\item{file_name_col}{name of column that comprises the texts file names}

\item{path_results}{path to directory where results should be saved}
}
\description{
This function creates subsets from metadata tables by detecting unique
values and using these values to split up the dataset.
}
\details{
To enable the combination with zscores and/or distance tables, the metadata
tables have to feature a column featuring the respective texts filename.
}
\examples{
filterSubsets()
}
\keyword{subsets}
