% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestNeighbours.R
\name{getNearestNeighbours}
\alias{getNearestNeighbours}
\title{Get Nearest Neighbours}
\usage{
getNearestNeighbours(dist, num)
}
\arguments{
\item{dist}{distance matrix}

\item{num}{number of nearest neighbours}
}
\description{
This function leaves only values that belong to the x nearest neighours in
a distance matrix, all other values are replaced by 0.
}
\examples{
getNearestNeighbours()
}
\keyword{nearest}
\keyword{neighbours}
