% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{ngram}
\alias{ngram}
\title{Transform a Tokens Object to n-grams}
\usage{
ngram(preprocessed_corpus, n_gram_size = 1)
}
\arguments{
\item{preprocessed_corpus}{a quanteda tokens object containing all tokens
in the corpus}
}
\description{
This function returns either uni- or bigrams.
}
\examples{
ngram()
}
\keyword{ngram}
