% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseZscores.R
\name{normaliseZscores}
\alias{normaliseZscores}
\title{Normalise Zscores}
\usage{
normaliseZscores(zscores)
}
\arguments{
\item{zscores}{matrix of zscores}
}
\description{
This function transforms zscores to normalised zscores.
}
\examples{
normaliseZscores()
}
\keyword{transformation}
\keyword{zscore}
