% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performCulling.R
\name{performCulling}
\alias{performCulling}
\title{Perform Culling}
\usage{
performCulling(freq_dist, culling)
}
\arguments{
\item{freq_dist}{frequency distribution matrix}

\item{culling}{culling value}
}
\description{
This function performs culling and returns the culled frequency distribution.
}
\examples{
performCulling()
}
\keyword{culling}
