% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortOverallFeatureList.R
\name{sortOverallFeatureList}
\alias{sortOverallFeatureList}
\title{Sort an Overall Feature List}
\usage{
sortOverallFeatureList(n_grams)
}
\arguments{
\item{n_grams}{a quanteda tokens object containing all tokens
in the corpus turned into n-grams}
}
\description{
This function creates and sorts an feature list for the entire corpus.
}
\examples{
sortOverallFeatureList()
}
\keyword{feature}
\keyword{list}
\keyword{sorted}
