% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternariseZscores.R
\name{ternariseZscores}
\alias{ternariseZscores}
\title{Ternarise Zscores}
\usage{
ternariseZscores(zscores)
}
\arguments{
\item{zscores}{matrix of zscores}
}
\description{
This function transforms zscores to ternarised zscores.
}
\examples{
ternariseZscores()
}
\keyword{transformation}
\keyword{zscore}
