% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSignificance.R
\name{testSignificance}
\alias{testSignificance}
\title{Test the Significance}
\usage{
testSignificance(path_dist_1, path_dist_2, stat_value)
}
\arguments{
\item{path_dist_1}{path to distance table comprising texts of category A}

\item{stat_value}{statistical value that will be used for the comparison,
possible values: mean, median, standard_deviation}

\item{path_dist_1}{path to distance table comprising texts of category B}
}
\description{
This function tests the significance of correlations between values from
two subsets.
}
\examples{
testSignificance()
}
\keyword{significance}
